Source Folder v4
made the meat of the code into a source, while making the code for the player to start with be the main code.

Bit more organized on the backend / front end for the code.

Backend will have the random / handling on how the cards being delt with. While the front end will be for player code