updated card game for future reference v5
Updated the program for future reference. Feel free to test it out.

To start the game play it from main_card.
The rest of the info can be found in comments in regarding on how the game works.
Its basically a card game, after seeing how its built from Cpt. Jimmy Kirk
Ive added some new features to it to help brighten it up a bit.

Will eventually make this into a gui, for now will finish up on learning a bit more for web interface by using
python.

Video for reference:
https://www.youtube.com/watch?v=62TmpPDs0mM